pyproject.toml
[tool.poetry]
name = "job-analyzer"
version = "0.1.0"
description = "Multi-agent AWS Job Description Analyzer"
authors = ["Mani <sattumanisharathchandra@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.34.0"
aws-cdk-lib = "^2.120.0"
constructs = "^10.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-mock = "^3.12.0"
moto = {extras = ["dynamodb", "s3", "lambda"], version = "^4.2.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"